+++ tests/game/effects/clock.test.ts
@@ -6,7 +6,7 @@ import query from 'common/components/query'
 import {GameModel} from 'common/models/game-model'
 import TurnSkippedEffect from 'common/status-effects/turn-skipped'
 import UsedClockEffect from 'common/status-effects/used-clock'
-import {applyEffect, endTurn, playCardFromHand, testGame} from './utils'
+import {applyEffect, endTurn, playCardFromHand, testGame} from '../utils'
 
 function* testClockHelperSaga(game: GameModel) {
 	yield* playCardFromHand(game, EthosLabCommon, 'hermit', 0)
